requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
altgraph==0.17.2
aniso8601==9.0.1
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
config==0.5.1
crcmod==1.7
Flask==2.0.3
flask-marshmallow==0.14.0
Flask-MQTT==1.1.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
future==0.18.2
greenlet==1.1.2
idna==3.3
ifaddr==0.1.7
iniconfig==1.1.1
iso8601==1.0.2
itsdangerous==2.1.0
Jinja2==3.0.3
lxml==4.8.0
MarkupSafe==2.1.0
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
packaging==21.3
paho-mqtt==1.6.1
pefile==2021.9.3
pluggy==1.0.0
py==1.11.0
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.2
pyparsing==3.0.7
pyserial==3.5
pyserial-asyncio==0.6
pytest==7.1.1
python-rapidjson==1.6
pytils==0.3
pytz==2021.3
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.27.1
six==1.16.0
sql==0.4.0
SQLAlchemy==1.4.31
tomli==2.0.1
tools==0.1.9
transitions==0.8.11
urllib3==1.26.9
Werkzeug==2.0.3
xmodem==0.4.6
zeroconf==0.38.4
